pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "securescan-ai"
version = "0.1.0"
description = "AI-powered security scanner with CVE intelligence"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Sai Mani", email = "psmk212004@gmail.com"}]
keywords = ["security", "sast", "ai", "llm", "vulnerability", "scanning", "cve"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "semgrep>=1.45.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
]

[project.urls]
Homepage = "https://github.com/saimani21/securescan-ai"
Repository = "https://github.com/saimani21/securescan-ai"
"Bug Tracker" = "https://github.com/saimani21/securescan-ai/issues"

[project.scripts]
secscan = "securescan.cli.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["securescan*"]
